[FIX] base: make tests independent of default user groups
The tests should not be impacted by the demo data however it is
currently the case.
Indeed, a lot of groups are added to the default group user when
installing the demo data.
Instead, we only take the user group like it is the case when no demo is
installed.

closes odoo/odoo#223039

Related: odoo/enterprise#92410
Signed-off-by: Chong Wang (cwg) <cwg@odoo.com>